Renamed 'parser' to 'dotapatch'
See #8
