+++ dotapatch/changelogs/688e
