+++ dotapatch/templates/dota2
